[BE/FIX] Q 클래스 생기도록 수정

+++ backend/src/main/java/com/prana/backend/yoga/repository/YogaRepository.java
@@ -4,6 +4,6 @@ package com.prana.backend.yoga.repository;
 import com.prana.backend.entity.Yoga;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface YogaRepository extends JpaRepository<Yoga, Long> {
+public interface YogaRepository extends JpaRepository<Yoga, Integer> {
 
 }
